+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-## TBD
+## 1.6.2 (2024-07-04)
 
 ### Bug fixes
 
@@ -9,7 +9,7 @@ Changelog
 
 ### Enhancements
 
-* Span parentage APIs now require a BugsnagPerformanceSpanContext, which BugsnagPerformanceSpan is now a subclass of. You no longer need to assign a Bugsnag span as a parent. [280](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/280)
+* Span parentage APIs now take a BugsnagPerformanceSpanContext, which BugsnagPerformanceSpan is now a subclass of. You no longer need to assign a BugsnagPerformanceSpan as a parent. [280](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/280)
 
 ## 1.6.1 (2024-06-24)